move gittest.BasicAuthMW to mwtest package
